requirements.txt

# Core web framework
fastapi>=0.95.0
uvicorn[standard]>=0.20.0

# Environment variable loader
python-dotenv>=1.0.0

# OpenAI API client
openai>=1.0.0

# Data validation
pydantic>=1.10.0

# Hugging Face Transformers for local model inference
transformers>=4.0.0
torch>=1.12.0
